User Interfaces Merge

+++ src/interactable.cpp
@@ -142,14 +142,13 @@ void Interactable::interact() {
 	if (Game::getBin().scripts.count(intBin.scriptPath) != 0) {
 		auto bc = Game::getBin().scripts[intBin.scriptPath].bytecode;
 		auto result = state.safe_script(bc, &sol::script_pass_on_error);
-		// auto unsafe_result = state.unsafe_script(bc);
 
 		if (!result.valid()) {
 			sol::error error = result;
 			std::cout << error.what() << std::endl;
 		}
 		if (result.status() != sol::call_status::ok) {
-			printf("uh oh: %i \n", result.status());
+			printf("uh oh: %i \n", static_cast<int>(result.status()));
 		}
 
 		if (state["interact"].valid()) {
